+++ test/test_mysql.js
@@ -12,6 +12,8 @@ describe(__filename, function () {
   it('connect error is detected during init()', async function () {
     // Use an invalid TCP port to force a connection error.
     const db = new mysql.Database({...databases.mysql, port: 65536});
+    // An error is expected; prevent it from being logged.
+    db.logger = Object.setPrototypeOf({error() {}}, db.logger);
     await assert.rejects(db.init());
   });
 
@@ -19,6 +21,8 @@ describe(__filename, function () {
     const db = new mysql.Database(databases.mysql);
     await db.init();
     const before = await db._connection;
+    // An error is expected; prevent it from being logged.
+    db.logger = Object.setPrototypeOf({error() {}}, db.logger);
     // Sleep longer than the timeout to force a fatal error.
     await assert.rejects(db._query({sql: 'DO SLEEP(1);', timeout: 1}), {fatal: true});
     const after = await db._connection;